# x/api/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
...
    AuthConfig:
        properties:
            password:
                type: string
                x-go-name: Password
            username:
                type: string
                x-go-name: Username
        type: object
        x-go-package: github.com/168yy/netx/x/config
...
    ChainConfig:
        properties:
            hops:
                items:
                    $ref: '#/definitions/HopConfig'
                type: array
                x-go-name: Hops
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            name:
                type: string
                x-go-name: Name
        type: object
        x-go-package: github.com/168yy/netx/x/config
...
    Duration:
        description: |-
            A Duration represents the elapsed time between two instants
            as an int64 nanosecond count. The representation limits the
            largest representable duration to approximately 290 years.
        format: int64
        type: integer
        x-go-package: time
    FileLoader:
        properties:
            path:
                type: string
                x-go-name: Path
        type: object
        x-go-package: github.com/168yy/netx/x/config
...
    ForwardNodeConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            auth:
                $ref: '#/definitions/AuthConfig'
            bypass:
                type: string
                x-go-name: Bypass
            bypasses:
                items:
                    type: string
                type: array
                x-go-name: Bypasses
            filter:
                $ref: '#/definitions/NodeFilterConfig'
            host:
                description: DEPRECATED by filter.host
                type: string
                x-go-name: Host
            http:
                $ref: '#/definitions/HTTPNodeConfig'
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            name:
                type: string
                x-go-name: Name
            network:
                type: string
                x-go-name: Network
            path:
                description: DEPRECATED by filter.path
                type: string
                x-go-name: Path
            protocol:
                description: DEPRECATED by filter.protocol
                type: string
                x-go-name: Protocol
            tls:
                $ref: '#/definitions/TLSNodeConfig'
        type: object
        x-go-package: github.com/168yy/netx/x/config
    ForwarderConfig:
        properties:
            hop:
                description: the referenced hop name
                type: string
                x-go-name: Hop
            name:
                description: DEPRECATED by hop field
                type: string
                x-go-name: Name
            nodes:
                items:
                    $ref: '#/definitions/ForwardNodeConfig'
                type: array
                x-go-name: Nodes
            selector:
                $ref: '#/definitions/SelectorConfig'
        type: object
        x-go-package: github.com/168yy/netx/x/config
    HTTPLoader:
        properties:
            timeout:
                $ref: '#/definitions/Duration'
            url:
                type: string
                x-go-name: URL
        type: object
        x-go-package: github.com/168yy/netx/x/config
...
    HandlerConfig:
        properties:
            auth:
                $ref: '#/definitions/AuthConfig'
            auther:
                type: string
                x-go-name: Auther
            authers:
                items:
                    type: string
                type: array
                x-go-name: Authers
            chain:
                type: string
                x-go-name: Chain
            chainGroup:
                $ref: '#/definitions/ChainGroupConfig'
            limiter:
                type: string
                x-go-name: Limiter
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            observer:
                type: string
                x-go-name: Observer
            retries:
                format: int64
                type: integer
                x-go-name: Retries
            tls:
                $ref: '#/definitions/TLSConfig'
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: github.com/168yy/netx/x/config
    HopConfig:
        properties:
            bypass:
                type: string
                x-go-name: Bypass
            bypasses:
                items:
                    type: string
                type: array
                x-go-name: Bypasses
            file:
                $ref: '#/definitions/FileLoader'
            hosts:
                type: string
                x-go-name: Hosts
            http:
                $ref: '#/definitions/HTTPLoader'
            interface:
                type: string
                x-go-name: Interface
            name:
                type: string
                x-go-name: Name
            nodes:
                items:
                    $ref: '#/definitions/NodeConfig'
                type: array
                x-go-name: Nodes
            plugin:
                $ref: '#/definitions/PluginConfig'
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
            resolver:
                type: string
                x-go-name: Resolver
            selector:
                $ref: '#/definitions/SelectorConfig'
            sockopts:
                $ref: '#/definitions/SockOptsConfig'
        type: object
        x-go-package: github.com/168yy/netx/x/config
...
    ListenerConfig:
        properties:
            auth:
                $ref: '#/definitions/AuthConfig'
            auther:
                type: string
                x-go-name: Auther
            authers:
                items:
                    type: string
                type: array
                x-go-name: Authers
            chain:
                type: string
                x-go-name: Chain
            chainGroup:
                $ref: '#/definitions/ChainGroupConfig'
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            tls:
                $ref: '#/definitions/TLSConfig'
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: github.com/168yy/netx/x/config
...
    NodeConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            bypass:
                type: string
                x-go-name: Bypass
            bypasses:
                items:
                    type: string
                type: array
                x-go-name: Bypasses
            connector:
                $ref: '#/definitions/ConnectorConfig'
            dialer:
                $ref: '#/definitions/DialerConfig'
            filter:
                $ref: '#/definitions/NodeFilterConfig'
            hosts:
                type: string
                x-go-name: Hosts
            http:
                $ref: '#/definitions/HTTPNodeConfig'
            interface:
                type: string
                x-go-name: Interface
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            name:
                type: string
                x-go-name: Name
            network:
                type: string
                x-go-name: Network
            resolver:
                type: string
                x-go-name: Resolver
            sockopts:
                $ref: '#/definitions/SockOptsConfig'
            tls:
                $ref: '#/definitions/TLSNodeConfig'
        type: object
        x-go-package: github.com/168yy/netx/x/config
...
    PluginConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            timeout:
                $ref: '#/definitions/Duration'
            tls:
                $ref: '#/definitions/TLSConfig'
            token:
                type: string
                x-go-name: Token
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: github.com/168yy/netx/x/config
...
    RecorderObject:
        properties:
            Metadata:
                additionalProperties: {}
                type: object
            name:
                type: string
                x-go-name: Name
            record:
                type: string
                x-go-name: Record
        type: object
        x-go-package: github.com/168yy/netx/x/config
    RedisLoader:
        properties:
            addr:
                type: string
                x-go-name: Addr
            db:
                format: int64
                type: integer
                x-go-name: DB
            key:
                type: string
                x-go-name: Key
            password:
                type: string
                x-go-name: Password
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: github.com/168yy/netx/x/config
...
    Response:
        properties:
            code:
                format: int64
                type: integer
                x-go-name: Code
            msg:
                type: string
                x-go-name: Msg
        type: object
        x-go-package: github.com/168yy/netx/x/api
    RouterConfig:
        properties:
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            name:
                type: string
                x-go-name: Name
            plugin:
                $ref: '#/definitions/PluginConfig'
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
            routes:
                items:
                    $ref: '#/definitions/RouterRouteConfig'
                type: array
                x-go-name: Routes
        type: object
        x-go-package: github.com/168yy/netx/x/config
    RouterRouteConfig:
        properties:
            gateway:
                type: string
                x-go-name: Gateway
            net:
                type: string
                x-go-name: Net
        type: object
        x-go-package: github.com/168yy/netx/x/config
...
    SelectorConfig:
        properties:
            failTimeout:
                $ref: '#/definitions/Duration'
            maxFails:
                format: int64
                type: integer
                x-go-name: MaxFails
            strategy:
                type: string
                x-go-name: Strategy
        type: object
        x-go-package: github.com/168yy/netx/x/config
    ServiceConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            admission:
                type: string
                x-go-name: Admission
            admissions:
                items:
                    type: string
                type: array
                x-go-name: Admissions
            bypass:
                type: string
                x-go-name: Bypass
            bypasses:
                items:
                    type: string
                type: array
                x-go-name: Bypasses
            climiter:
                type: string
                x-go-name: CLimiter
            forwarder:
                $ref: '#/definitions/ForwarderConfig'
            handler:
                $ref: '#/definitions/HandlerConfig'
            hosts:
                type: string
                x-go-name: Hosts
            interface:
                description: DEPRECATED by metadata.interface since beta.5
                type: string
                x-go-name: Interface
            limiter:
                type: string
                x-go-name: Limiter
            listener:
                $ref: '#/definitions/ListenerConfig'
            logger:
                type: string
                x-go-name: Logger
            loggers:
                items:
                    type: string
                type: array
                x-go-name: Loggers
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            name:
                type: string
                x-go-name: Name
            observer:
                type: string
                x-go-name: Observer
            recorders:
                items:
                    $ref: '#/definitions/RecorderObject'
                type: array
                x-go-name: Recorders
            resolver:
                type: string
                x-go-name: Resolver
            rlimiter:
                type: string
                x-go-name: RLimiter
            sockopts:
                $ref: '#/definitions/SockOptsConfig'
            status:
                $ref: '#/definitions/ServiceStatus'
        type: object
        x-go-package: github.com/168yy/netx/x/config
...
    ServiceStatus:
        properties:
            createTime:
                format: int64
                type: integer
                x-go-name: CreateTime
            events:
                items:
                    $ref: '#/definitions/ServiceEvent'
                type: array
                x-go-name: Events
            state:
                type: string
                x-go-name: State
            stats:
                $ref: '#/definitions/ServiceStats'
        type: object
        x-go-package: github.com/168yy/netx/x/config
    SockOptsConfig:
        properties:
            mark:
                format: int64
                type: integer
                x-go-name: Mark
        type: object
        x-go-package: github.com/168yy/netx/x/config
...
    TLSConfig:
        properties:
            caFile:
                type: string
                x-go-name: CAFile
            certFile:
                type: string
                x-go-name: CertFile
            commonName:
                type: string
                x-go-name: CommonName
            keyFile:
                type: string
                x-go-name: KeyFile
            options:
                $ref: '#/definitions/TLSOptions'
            organization:
                type: string
                x-go-name: Organization
            secure:
                type: boolean
                x-go-name: Secure
            serverName:
                type: string
                x-go-name: ServerName
            validity:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: github.com/168yy/netx/x/config
...
info:
    title: Documentation of Web API.
    version: 1.0.0
paths:
...
    /config/chains:
        post:
            operationId: createChainRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/ChainConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createChainResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new chain, the name of chain must be unique in chain list.
            tags:
                - Chain
    /config/chains/{chain}:
        delete:
            operationId: deleteChainRequest
            parameters:
                - in: path
                  name: chain
                  required: true
                  type: string
                  x-go-name: Chain
            responses:
                "200":
                    $ref: '#/responses/deleteChainResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete chain by name.
            tags:
                - Chain
        put:
            operationId: updateChainRequest
            parameters:
                - in: path
                  name: chain
                  required: true
                  type: string
                  x-go-name: Chain
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/ChainConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateChainResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update chain by name, the chain must already exist.
            tags:
                - Chain
...
    /config/routers:
        post:
            operationId: createRouterRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/RouterConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createRouterResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new router, the name of the router must be unique in router list.
            tags:
                - Router
    /config/routers/{router}:
        delete:
            operationId: deleteRouterRequest
            parameters:
                - in: path
                  name: router
                  required: true
                  type: string
                  x-go-name: Router
            responses:
                "200":
                    $ref: '#/responses/deleteRouterResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete router by name.
            tags:
                - Router
        put:
            operationId: updateRouterRequest
            parameters:
                - in: path
                  name: router
                  required: true
                  type: string
                  x-go-name: Router
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/RouterConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateRouterResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update router by name, the router must already exist.
            tags:
                - Router
...
    /config/services:
        post:
            operationId: createServiceRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/ServiceConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createServiceResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new service, the name of the service must be unique in service list.
            tags:
                - Service
    /config/services/{service}:
        delete:
            operationId: deleteServiceRequest
            parameters:
                - in: path
                  name: service
                  required: true
                  type: string
                  x-go-name: Service
            responses:
                "200":
                    $ref: '#/responses/deleteServiceResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete service by name.
            tags:
                - Service
        put:
            operationId: updateServiceRequest
            parameters:
                - in: path
                  name: service
                  required: true
                  type: string
                  x-go-name: Service
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/ServiceConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateServiceResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update service by name, the service must already exist.
            tags:
                - Service
produces:
    - application/json
responses:
...
    createChainResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    createRouterResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    createServiceResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    deleteChainResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    deleteRouterResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    deleteServiceResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    updateChainResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    updateRouterResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    updateServiceResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
schemes:
    - https
    - http
securityDefinitions:
    basicAuth:
        type: basic
swagger: "2.0"